+++ CHANGELOG.md
@@ -7,7 +7,7 @@ All Notable changes to `League\Uri\Interfaces` will be documented in this file
 ### Added
 
 - `QueryString::compose` and `QueryString::composeFromValue` a userland improved `http_build_query` method.
-- `QueryParsingMode` and `QueryBuildingMode` Enum to improve query parsing and building.
+- `QueryExtractMode` and `QueryComposeMode` Enum to improve query parsing and building.
 
 ### Fixed